aula06/requirements.txt
astroid==2.3.1
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.3
pylint==2.4.2
six==1.12.0
SQLAlchemy==1.3.10
typed-ast==1.4.0
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
